Change ArrayTypesLocation
