+++ src/types/ArrayProps.ts
@@ -1,0 +1,10 @@
+export interface ArrayProps {
+  isYes: boolean;
+  isFinal: boolean;
+  Playing: string[];
+  TeamName: string[];
+  Grade: string[];
+  Time: string[];
+  isLive: boolean;
+  isVoting: boolean;
+}
